[REFACTOR]: Add useVideo customHook
