+++ src/hooks/useVideo.js
@@ -1,0 +1,119 @@
+import React, { useEffect, useRef, useState } from "react";
+import Peer from "simple-peer";
+
+import { socket, socketApi } from "../utils/socket";
+
+export default function useVideo() {
+  const [peers, setPeers] = useState([]);
+  const userVideo = useRef();
+  const peersRef = useRef([]);
+  const [itUser, setItUser] = useState(null);
+  const [participantUser, setParticipantUser] = useState(null);
+  const [isRedadyPoseDetection, setIsReadyPoseDetection] = useState(false);
+  const [difficulty, setDifficulty] = useState(null);
+
+  const videoConstraints = {
+    height: window.innerHeight / 2,
+    width: window.innerWidth / 2,
+  };
+
+  useEffect(() => {
+    navigator.mediaDevices
+      .getUserMedia({
+        video: videoConstraints,
+        audio: true,
+      })
+      .then((stream) => {
+        if (userVideo.current) {
+          userVideo.current.srcObject = stream;
+        }
+
+        socketApi.enterGameRoom(true);
+
+        socket.on("all-info", (payload) => {
+          console.log(payload);
+          setItUser(payload.it);
+          setParticipantUser(payload.participant);
+          setDifficulty(payload.difficulty);
+          setIsReadyPoseDetection(true);
+
+          const peers = [];
+
+          payload.socketInRoom.forEach((user) => {
+            const peer = new Peer({
+              initiator: true,
+              trickle: false,
+              stream,
+            });
+
+            peer.on("signal", (signal) => {
+              console.log("this is signal", signal);
+              socket.emit("sending signal", {
+                userToSignal: user,
+                callerID: socket.id,
+                signal,
+              });
+            });
+
+            peersRef.current.push({
+              peerID: user,
+              peer,
+            });
+            peers.push(peer);
+          });
+          setPeers(peers);
+        });
+
+        socket.on("user joined", (payload) => {
+          const peer = addPeer(payload.signal, payload.callerID, stream);
+          peersRef.current.push({
+            peerID: payload.callerID,
+            peer,
+          });
+
+          setPeers((users) => [...users, peer]);
+        });
+
+        socket.on("receiving-returned-signal", (payload) => {
+          const item = peersRef.current.find((p) => p.peerID === payload.id);
+          item.peer.signal(payload.signal);
+        });
+      });
+
+    return () => {
+      userVideo.current = null;
+
+      socket.off("all-info");
+      socket.off("user joined");
+      socket.off("receiving-returned-signal");
+    };
+  }, []);
+
+  function addPeer(incomingSignal, callerID, stream) {
+    const peer = new Peer({
+      initiator: false,
+      trickle: false,
+      stream,
+    });
+
+    peer.on("signal", (signal) => {
+      console.log(signal, "누가 들어왓대", callerID, "<-얘가 왔대");
+      socket.emit("returning signal", { signal, callerID });
+    });
+    console.log("this is incomingSignal", incomingSignal);
+    peer.signal(incomingSignal);
+
+    return peer;
+  }
+
+  return {
+    userVideo,
+    participantUser,
+    itUser,
+    difficulty,
+    isRedadyPoseDetection,
+    peers,
+    peersRef,
+    setParticipantUser,
+  };
+}
